Task-13 Элемент списка вакансий Вынес функцию formatSalary в отдельный object UtilityFunctions
